graft: Add tls module
