+++ graft/src/sonar.rs
@@ -12,20 +12,15 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
 //
+use crate::tls;
 use http::StatusCode;
-use rustls::client::danger::{
-    HandshakeSignatureValid, ServerCertVerified, ServerCertVerifier,
-};
-use rustls::pki_types::{CertificateDer, ServerName, UnixTime};
-use rustls::{DigitallySignedStruct, SignatureScheme};
+use rustls::pki_types::ServerName;
 use std::net::ToSocketAddrs;
-use std::sync::Arc;
 use std::time::Duration;
 use tokio::io::{self, AsyncReadExt, AsyncWriteExt, ErrorKind};
 use tokio::net::TcpStream;
 use tokio::time::timeout;
 use tokio_rustls::client::TlsStream;
-use tokio_rustls::TlsConnector;
 
 /// Sonar protocol error
 #[derive(Debug, thiserror::Error)]
@@ -83,62 +78,6 @@ impl From<Error> for StatusCode {
 
 /// Sonar result
 pub type Result<T> = std::result::Result<T, Error>;
-
-/// TLS certificate verifier for IRIS server
-#[derive(Debug)]
-struct Verifier {}
-
-impl ServerCertVerifier for Verifier {
-    fn verify_server_cert(
-        &self,
-        _: &CertificateDer,
-        _: &[CertificateDer],
-        _: &ServerName,
-        _: &[u8],
-        _: UnixTime,
-    ) -> std::result::Result<ServerCertVerified, rustls::Error> {
-        // NOTE: we are *NOT* checking the server cert here!  This is only
-        //       secure when graft is running on the same host as IRIS.
-        Ok(ServerCertVerified::assertion())
-    }
-
-    fn verify_tls12_signature(
-        &self,
-        _: &[u8],
-        _: &CertificateDer<'_>,
-        _: &DigitallySignedStruct,
-    ) -> std::result::Result<HandshakeSignatureValid, rustls::Error> {
-        Ok(HandshakeSignatureValid::assertion())
-    }
-
-    fn verify_tls13_signature(
-        &self,
-        _: &[u8],
-        _: &CertificateDer<'_>,
-        _: &DigitallySignedStruct,
-    ) -> std::result::Result<HandshakeSignatureValid, rustls::Error> {
-        Ok(HandshakeSignatureValid::assertion())
-    }
-
-    fn supported_verify_schemes(&self) -> Vec<SignatureScheme> {
-        use SignatureScheme::*;
-        [
-            RSA_PKCS1_SHA1,
-            RSA_PKCS1_SHA256,
-            ECDSA_NISTP256_SHA256,
-            RSA_PKCS1_SHA384,
-            ECDSA_NISTP384_SHA384,
-            RSA_PKCS1_SHA512,
-            ECDSA_NISTP521_SHA512,
-            RSA_PSS_SHA256,
-            RSA_PSS_SHA384,
-            RSA_PSS_SHA512,
-            ED25519,
-            ED448,
-        ]
-        .into()
-    }
-}
 
 impl Error {
     /// Parse a SHOW message received from server
@@ -326,11 +265,7 @@ impl Connection {
             .next()
             .ok_or_else(|| io::Error::from(ErrorKind::NotFound))?;
         let tcp_stream = TcpStream::connect(&addr).await?;
-        let cfg = rustls::ClientConfig::builder()
-            .dangerous()
-            .with_custom_certificate_verifier(Arc::new(Verifier {}))
-            .with_no_client_auth();
-        let connector = TlsConnector::from(Arc::new(cfg));
+        let connector = tls::connector();
         let domain = ServerName::try_from(host)
             .map_err(|_| {
                 io::Error::new(io::ErrorKind::InvalidInput, "invalid dnsname")
